diff-service/requirements.txt
flask>=3.0,<4.0
flask-cors>=4.0,<5.0
pdfplumber>=0.11,<1.0
python-dotenv>=1.0,<2.0
